scripts to automatically find F0-range and power-threshold from histogram; 1.00 RT (cyclevae+mwdlp) achieved with 16-dimension middle-layer before final FC layer instead of 32-dim., and provide further stability; tested functions for ana-syn/convert with mel-spec output;

+++ demo_sparse-cyclevae_mwdlp-10bit/inc/wave.h
@@ -40,5 +40,5 @@ struct HEADER {
     so you need to convert the bytes to big-endian [MSB in smallest address] in code for the values to make sense.
 */
 
-int read_write_wav(FILE *fin, FILE *fout, short *num_reflected_right_edge_samples, long *num_samples, long *size_of_each_sample);
+short read_write_wav(FILE *fin, FILE *fout, short *num_reflected_right_edge_samples, long *num_samples, long *size_of_each_sample);
 long read_feat_write_wav(FILE* fin, FILE* fout, int bin_flag);
